requirements.txt
requests
pandas
yfinance
schedule
python-dotenv
pyyaml
rich
